requirements.txt
Django>=6,<6.1
wagtail==7.3.1
wagtail-localize>=1.0.0
polib>=1.2.0
django-filter>=24.0,<25
djangorestframework-simplejwt>=5.4,<6
django-cors-headers>=4.6,<5
drf-spectacular>=0.28,<1
reportlab>=4.0,<5
openpyxl>=3.1,<4
celery[redis]>=5.4,<6
django-celery-results>=2.5,<3
django-redis>=5.4,<6
gunicorn>=23.0,<24
psycopg2-binary>=2.9,<3
dj-database-url>=2.0,<3
